refactor(entity): editing id type. form Long to Integer.

+++ src/main/java/com/bit/healthpartnerboot/dto/MemberDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class MemberDTO {
-    private long seq;
+    private Integer seq;
     private String email;
     private String password;
     private String name;
